Add error region calculation for transmitters.
This basically just combines error information about the drone
position, and the LOB. The main limmitation right now is that
it doesn't take into account the covariances between the LOB and
either of the components of the position. This is not a
deal-breaker, however, because since the position and LOB are
measured by two entirely separate sensor systems, the covariance
between them should theoretically be close to zero.