Fix LOB selection case where it passes the source.
Before, if it determined once that the LOB needed to flip, it
would flip it always. This might not have always been the right
thing to do if the drone had passed by the transmitter. Now, it
actually checks for that case and handles it properly.

This includes a test as well.